review_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai>=1.40.0
pydantic>=2.8.2
python-multipart>=0.0.9
Pillow>=10.4.0
python-dotenv>=1.0.1